+++ 1-sprint-mission/src/main/java/com/sprint/mission/discodeit/repository/file/FileMessageRepository.java
@@ -2,18 +2,20 @@ package com.sprint.mission.discodeit.repository.file;
 
 import com.sprint.mission.discodeit.entity.Message;
 import com.sprint.mission.discodeit.entity.User;
+import org.springframework.stereotype.Repository;
 
 import java.io.*;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.UUID;
 
+@Repository
 public class FileMessageRepository implements FileMessageRepositoryInterface{
-    private static final FileMessageRepository INSTANCE = new FileMessageRepository();
-    private FileMessageRepository(){}
-    public static FileMessageRepository getInstance() {
-        return INSTANCE;
-    }
+//    private static final FileMessageRepository INSTANCE = new FileMessageRepository();
+//    private FileMessageRepository(){}
+//    public static FileMessageRepository getInstance() {
+//        return INSTANCE;
+//    }
 
     public Message createMessage(String text, UUID userId, String userName ,UUID channelId,String channelName) throws IOException, ClassNotFoundException {
         Message message = new Message();
