+++ 1-sprint-mission/src/main/resources/application.yml
@@ -1,0 +1,1 @@
+spring.application.name=discodeit
